requirements.txt
cycler==0.11.0
DateTime==4.4
docutils==0.18.1
fonttools==4.29.1
git+https://github.com/graphsense/graphsense-python.git
kiwisolver==1.3.2
matplotlib==3.5.1
numpy==1.22.2
packaging==21.3
Pillow==9.0.1
pymongo==4.0.1
pyparsing==3.0.7
python-dateutil==2.8.2
pytz==2021.3
six==1.16.0
statistics==1.0.3.5
tqdm==4.62.3
urllib3==1.26.8
zope.interface==5.4.0
python-dotenv==0.4.0
dotenv-cli==2.2.0
networkx==2.7.1

